+++ cli/aux/aux.h
@@ -6,6 +6,10 @@
 
 #define PROGRAM     "tec"
 
+#define ARRAY_SIZE(x) (sizeof(x) / sizeof((x)[0]))
+#define xstr(s) str(s)
+#define str(s) #s
+
 // TODO: Get rid of it (set options with default boolean values)
 #define NONEBOOL        -1      /* Not yet set boolean value */